fix(table): fix too many dynamic stylesheets
A first attempt at fixing it. Possibly a "fulhack". Top is set in `Thead` and `Tr` which have access
to stickyOffset, and then inherited in the `td` and `th`. The problem is that `thead` and `tr` can't
be sticky because of a Chrome bug, and `Thead` and `Tr` can't pass `stickyOffset` as a prop because
they're not guaranteed to be direct children.