Removed deprecated components and props (#511)
* feat: remove deprecated components

Removes `pane` and `segmented-control`, and the `isOpen` prop of the `popup-menu`.

Pane, SegmentedControl and the isOpen prop of PopupMenu is no longer available.

* test: update tests for removed deprecations

* feat: add deprecations and remove more components

Deprecate CostInfoLogo, InstrumentBadge and Tabs. Remove Dropdown and TableContentWrapper.

Dropdown and TableContentWrapper are no longer provided. Use PopupMenu instead for Dropdown.

* feat: remove CostInfoLogo too

CostInfoLogo no longer is provided by this project.

* feat: remove deprecated flag with numeric size property

- Removes possibility to pass `size` as a numeric value to Flag component. Use one of literal values (xs, sm, md, lg) instead.
- Removes possibility to pass properties `hideBorder` and `borderColor` since these can be passed together with `styles` prop.

* test: remove test for deprecated flag